requirements.txt
# Core dependencies
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.20.0

# SQL parsing and database
sqlparse>=0.4.4
pymysql>=1.1.0
cryptography>=41.0.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Code quality
ruff>=0.3.0

# v1.3.0 - Multi-provider support (optional)
# Uncomment based on your needs:

# AWS Bedrock (Claude models)
boto3>=1.34.0
botocore>=1.34.0

# Google GenAI (Gemini models)
google-genai>=1.39.1

# Retry mechanism and reliability
tenacity>=8.2.0

